Added init argument to lmrob() and lmrob.fit().
The init argument of lmrob can be a string, a function or a list. This is
evaluated in lmrob() and then passed to lmrob.fit(), which only takes a
list. lmrob.fit() is backwards compatible: if no init argument is given,
lmrob.S() is called as it was called before.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/robustbase@287 694ef91d-65df-0310-b7bb-92e67a308ead