New option cov.corrfact = "tauold" in .vcov.w
This option corresponds to the old cov.corrfact = "tau" option.
The new cov.corrfact = "tau" (and "hybrid") use
    corrfact <- mean(tau^2 * r.psi^2) / mpp2 * hcorr
where as "tauold" still uses
    corrfact <- mean(r.psi^2) / mpp2 * hcorr
which does not reduce to the classical solution.

To make cov.corrfact = "tau" and "hybrid" reduce to the classical solution,
the default for cov.dfcorr was changed to 1 (= mean(w) /  mean(tau^2)). It used
to be -1 (= mean(w)).

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/robustbase@565 694ef91d-65df-0310-b7bb-92e67a308ead